jrunscript.shell: remove invocation.js
Incorporate functionality into module.js to make it easier to organize
further
